Make user stack a uint64_t for convenience

+++ include/truth/process.h
@@ -21,9 +21,9 @@ struct thread {
     enum thread_state state;
     size_t user_stack_size;
     size_t kernel_stack_size;
-    uint8_t *user_stack;
-    uint8_t *user_stack_pointer;
-    uint8_t *kernel_stack;
+    uint64_t *user_stack;
+    uint64_t *user_stack_pointer;
+    uint64_t *kernel_stack;
     struct thread *next;
     struct thread *prev;
     struct process *process;
